+++ src/plugins/check_in/plugins/check_in_dietary.py
@@ -1,5 +1,7 @@
 from nonebot.plugin import PluginMetadata, inherit_supported_adapters
 from nonebot_plugin_alconna import Alconna, Args, CommandMeta, Match, on_alconna
+from nonebot_plugin_alconna.builtins.extensions.discord import DiscordSlashExtension
+from nonebot_plugin_alconna.builtins.extensions.telegram import TelegramSlashExtension
 from nonebot_plugin_user import UserSession
 
 from src.plugins.check_in.models import DietaryRecord
@@ -16,16 +18,20 @@ __plugin_meta__ = PluginMetadata(
 
 dietary_cmd = on_alconna(
     Alconna(
-        "饮食打卡",
+        "checkin_dietary",
         Args["content?#A：健康饮食少油少糖 B：我摆了偷吃了炸鸡奶茶", str],
         meta=CommandMeta(
             description=__plugin_meta__.description,
             example=__plugin_meta__.usage,
         ),
     ),
-    aliases={("check_in.dietary")},
+    aliases={"饮食打卡"},
     use_cmd_start=True,
     block=True,
+    extensions=[
+        TelegramSlashExtension(),
+        DiscordSlashExtension(name_localizations={"zh-CN": "饮食打卡"}),
+    ],
 )