+++ src/plugins/check_in/plugins/check_in_weight.py
@@ -1,5 +1,7 @@
 from nonebot.plugin import PluginMetadata, inherit_supported_adapters
 from nonebot_plugin_alconna import Alconna, Args, CommandMeta, Match, on_alconna
+from nonebot_plugin_alconna.builtins.extensions.discord import DiscordSlashExtension
+from nonebot_plugin_alconna.builtins.extensions.telegram import TelegramSlashExtension
 from nonebot_plugin_user import UserSession
 
 from src.plugins.check_in.models import WeightRecord
@@ -21,16 +23,20 @@ __plugin_meta__ = PluginMetadata(
 
 target_weight_cmd = on_alconna(
     Alconna(
-        "目标体重",
+        "target_weight",
         Args["target?#目标体重", str],
         meta=CommandMeta(
             description="设置和查看目标体重",
             example="查看目标体重\n/目标体重\n设置目标体重\n/目标体重 60",
         ),
     ),
-    aliases={("check_in.target_weight")},
+    aliases={"目标体重"},
     use_cmd_start=True,
     block=True,
+    extensions=[
+        TelegramSlashExtension(),
+        DiscordSlashExtension(name_localizations={"zh-CN": "目标体重"}),
+    ],
 )
 
 
@@ -70,16 +76,20 @@ async def _(session: AsyncSession, user: UserSession, target: str):
 
 weight_record_cmd = on_alconna(
     Alconna(
-        "体重打卡",
+        "checkin_weight",
         Args["content?#体重（kg）", str],
         meta=CommandMeta(
             description="记录体重",
             example="记录体重\n/体重打卡\n/体重打卡 60",
         ),
     ),
-    aliases={"记录体重", "check_in.weight_record"},
+    aliases={"体重打卡", "记录体重"},
     use_cmd_start=True,
     block=True,
+    extensions=[
+        TelegramSlashExtension(),
+        DiscordSlashExtension(name_localizations={"zh-CN": "体重打卡"}),
+    ],
 )