+++ src/plugins/ff14/plugins/ff14_fflogs/__init__.py
@@ -8,6 +8,8 @@ from typing import Literal, cast
 from nonebot import get_driver, logger
 from nonebot.plugin import PluginMetadata, inherit_supported_adapters
 from nonebot_plugin_alconna import Alconna, Args, At, CommandMeta, MultiVar, Text, on_alconna
+from nonebot_plugin_alconna.builtins.extensions.discord import DiscordSlashExtension
+from nonebot_plugin_alconna.builtins.extensions.telegram import TelegramSlashExtension
 from nonebot_plugin_datastore import get_plugin_data
 from nonebot_plugin_user import UserSession, get_user
 
@@ -54,6 +56,10 @@ fflogs_cmd = on_alconna(
     ),
     use_cmd_start=True,
     block=True,
+    extensions=[
+        TelegramSlashExtension(),
+        DiscordSlashExtension(),
+    ],
 )