+++ src/plugins/ff14/plugins/ff14_gate/__init__.py
@@ -4,6 +4,8 @@ from typing import Literal, cast
 
 from nonebot.plugin import PluginMetadata, inherit_supported_adapters
 from nonebot_plugin_alconna import Alconna, Args, CommandMeta, Match, on_alconna
+from nonebot_plugin_alconna.builtins.extensions.discord import DiscordSlashExtension
+from nonebot_plugin_alconna.builtins.extensions.telegram import TelegramSlashExtension
 
 from .data_source import get_direction
 
@@ -25,8 +27,13 @@ gate_cmd = on_alconna(
             example=__plugin_meta__.usage,
         ),
     ),
+    aliases={"藏宝选门", "选门"},
     use_cmd_start=True,
     block=True,
+    extensions=[
+        TelegramSlashExtension(),
+        DiscordSlashExtension(name_localizations={"zh-CN": "藏宝选门"}),
+    ],
 )