+++ src/plugins/morning/plugins/morning_greeting/__init__.py
@@ -5,6 +5,8 @@ from nonebot.log import logger
 from nonebot.params import Depends
 from nonebot.plugin import PluginMetadata, inherit_supported_adapters
 from nonebot_plugin_alconna import Alconna, Args, CommandMeta, Match, on_alconna
+from nonebot_plugin_alconna.builtins.extensions.discord import DiscordSlashExtension
+from nonebot_plugin_alconna.builtins.extensions.telegram import TelegramSlashExtension
 from nonebot_plugin_apscheduler import scheduler
 from nonebot_plugin_orm import get_session
 from nonebot_plugin_saa import PlatformTarget, Text, get_target
@@ -64,17 +66,21 @@ async def morning():
 
 morning_cmd = on_alconna(
     Alconna(
-        "早安",
+        "morning",
         Args["arg?#功能选项（on/off/update/today）", str],
         meta=CommandMeta(
             description=__plugin_meta__.description,
             example=__plugin_meta__.usage,
         ),
     ),
-    aliases={"morning"},
+    aliases={"早安"},
     use_cmd_start=True,
     block=True,
     permission=admin_permission(),
+    extensions=[
+        TelegramSlashExtension(),
+        DiscordSlashExtension(name_localizations={"zh-CN": "早安"}),
+    ],
 )