+++ src/plugins/roll/plugins/nga_roll/__init__.py
@@ -4,6 +4,8 @@ import re
 
 from nonebot.plugin import PluginMetadata, inherit_supported_adapters
 from nonebot_plugin_alconna import Alconna, Args, CommandMeta, Match, on_alconna
+from nonebot_plugin_alconna.builtins.extensions.discord import DiscordSlashExtension
+from nonebot_plugin_alconna.builtins.extensions.telegram import TelegramSlashExtension
 
 from .data_source import roll_dices
 
@@ -28,6 +30,10 @@ roll_cmd = on_alconna(
     ),
     use_cmd_start=True,
     block=True,
+    extensions=[
+        TelegramSlashExtension(),
+        DiscordSlashExtension(),
+    ],
 )