+++ tests/plugins/bihua/test_delete.py
@@ -7,7 +7,7 @@ from tests.fake import fake_group_message_event_v11
 
 async def test_bihua_delete_not_found(app: App):
     """测试删除壁画"""
-    from src.plugins.bihua import delete_bihua_cmd
+    from src.plugins.bihua import bihua_delete_cmd
 
     async with app.test_matcher() as ctx:
         adapter = get_adapter(Adapter)
@@ -19,4 +19,4 @@ async def test_bihua_delete_not_found(app: App):
         )
         ctx.receive_event(bot, event)
         ctx.should_call_send(event, "删除失败：壁画不存在", True)
-        ctx.should_finished(delete_bihua_cmd)
+        ctx.should_finished(bihua_delete_cmd)